+++ src/backend/outbuf.c
@@ -36,7 +36,7 @@ Outbuffer::Outbuffer()
     inc = 0;
 }
 
-Outbuffer::Outbuffer(unsigned bufinc)
+Outbuffer::Outbuffer(size_t bufinc)
 {
     buf = NULL;
     pend = NULL;
@@ -61,11 +61,11 @@ void Outbuffer::reset()
 }
 
 // Reserve nbytes in buffer
-void Outbuffer::reserve(unsigned nbytes)
+void Outbuffer::reserve(size_t nbytes)
 {
     if (pend - p < nbytes)
-    {   unsigned oldlen = len;
-        unsigned used = p - buf;
+    {   size_t oldlen = len;
+        size_t used = p - buf;
 
         if (inc > nbytes)
         {
@@ -106,9 +106,9 @@ void Outbuffer::reserve(unsigned nbytes)
 //
 //      If data will append to buffer
 //              position for write and return new size
-int Outbuffer::position(unsigned pos, unsigned nbytes)
+size_t Outbuffer::position(size_t pos, size_t nbytes)
 {
-    int current_sz = size();
+    size_t current_sz = size();
     unsigned char *fend = buf+pos+nbytes;       // future end of buffer
     if (fend >= pend)
     {
@@ -119,7 +119,7 @@ int Outbuffer::position(unsigned pos, unsigned nbytes)
 }
 
 // Write an array to the buffer.
-void Outbuffer::write(const void *b, int len)
+void Outbuffer::write(const void *b, size_t len)
 {
     if (pend - p < len)
         reserve(len);
@@ -128,7 +128,7 @@ void Outbuffer::write(const void *b, int len)
 }
 
 // Write n zeros to the buffer.
-void *Outbuffer::writezeros(unsigned len)
+void *Outbuffer::writezeros(size_t len)
 {
     if (pend - p < len)
         reserve(len);
@@ -229,7 +229,7 @@ void Outbuffer::prependBytes(const char *s)
     prepend(s, strlen(s));
 }
 
-void Outbuffer::prepend(const void *b, unsigned len)
+void Outbuffer::prepend(const void *b, size_t len)
 {
     reserve(len);
     memmove(buf + len,buf,p - buf);
@@ -265,7 +265,7 @@ char *Outbuffer::toString()
  * Set current size of buffer.
  */
 
-void Outbuffer::setsize(unsigned size)
+void Outbuffer::setsize(size_t size)
 {
     p = buf + size;
 }
